GUI improvements, added Tracking

+++ dl_objectsegmentator/Net/utils.py
@@ -385,7 +385,6 @@ def resize_image(image, min_dim=None, max_dim=None, padding=False):
     h, w = image.shape[:2]
     window = (0, 0, h, w)
     scale = 1
-
     # Scale?
     if min_dim:
         # Scale up but not down
@@ -394,11 +393,14 @@ def resize_image(image, min_dim=None, max_dim=None, padding=False):
     if max_dim:
         image_max = max(h, w)
         if round(image_max * scale) > max_dim:
-            scale = max_dim / image_max
+            scale = float(max_dim) / float(image_max)
+            #scale = max_dim / image_max
     # Resize image and mask
     if scale != 1:
         image = scipy.misc.imresize(
-            image, (round(h * scale), round(w * scale)))
+            image, (int(round(h * scale)), int(round(w * scale))))
+        #image = scipy.misc.imresize(
+            #image, (round(h * scale), round(w * scale)))
     # Need padding?
     if padding:
         # Get new height and width
